# ast: Generic/BeforeAfter/Var Visitor specific Walk implementations.