… ast/term_test.go
… ast/visit_test.go
… rego/rego.go
… repl/repl.go
… topdown/bindings.go
… topdown/eval.go