@koponen-styra 2020-01-17
I'll try improving this a bit further, if the additional complexity doesn't seem too bad.