@koponen-styra 2020-01-21
I thought about it. Unfortunately, I don't think GenericVisitor is capable of replacing walk built on Visitor (as it allows changing the Visitor during the walk).