@tsandall 2020-01-21
I see... 🤦‍♀ 

If we do this, it would be good to update all usage within OPA to rely on the new implementation and then put on a comment on the old implementation to indicate that it's deprecated. After a few releases we could remove the old interface.