@koponen-styra 2020-01-22
Converted all the uses to the new implementations.