@koponen-styra 2020-01-22
I just realized the merged version does actually change the Visitor interface, instead of merely deprecating it. If you guys think it's worthwhile, I can prepare another PR to keep the Visitor interface unchanged as none of the new stuff depends on it.