@tsandall 2020-01-23
whoops. I missed that in my last review. Thanks for fixing it. That said, I did a search across github and only found a couple instances where other projects are relying on the visitor interface so it's probably fairly impact.